% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{count_as_ap}
\alias{count_as_ap}
\title{Convert to AP Number}
\usage{
count_as_ap(value)
}
\arguments{
\item{value}{A single positive integer}
}
\value{
For numbers 1-9, returns the number spelled out. Otherwise, returns
the number as a string.
}
\description{
Convert to AP Number
}
\examples{
count_as_ap(3)
count_as_ap(20)
}
