% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{count_as_ordinal}
\alias{count_as_ordinal}
\title{Transform a count to an ordinal string}
\usage{
count_as_ordinal(value)
}
\arguments{
\item{value}{A single positive integer}
}
\value{
A string with the ordinal representation of a number
}
\description{
Transform a count to an ordinal string
}
\examples{
count_as_ordinal(1)
count_as_ordinal(111)
}
