% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{count_as_word}
\alias{count_as_word}
\title{Convert Large Counts into Friendly Text}
\usage{
count_as_word(value, fmt = "\%.1f")
}
\arguments{
\item{value}{A single positive integer}

\item{fmt}{Extra number formatting supplied to sprintf}
}
\value{
Returns a string with the power of a number replaced by the appropriate word.
}
\description{
Note - currently limited to \code{.Machine$integer.max}.
}
\examples{
count_as_word(100)
count_as_word(1000000)
count_as_word(1200000000)
}
