% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize.R
\name{humanize.now}
\alias{humanize.now}
\title{Humanize Time Periods}
\usage{
humanize.now(t)
}
\arguments{
\item{t}{A date or date/time pair of class POSIXct, POSIXlt, Date, or a string that can be coerced into these classes.}
}
\description{
This functions presents a time as its distance from the time the function was called, using the largest possible time period (seconds, minutes, hours, days, weeks, months, or years) that will generate an integer equal to or larger than one.
}
\examples{
humanize.now("1978-12-15 2:15:00 Asia/Jerusalem")
}
\keyword{date,}
\keyword{simplify}
\keyword{time,}
