% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{natural_date}
\alias{natural_date}
\title{Natural Date}
\usage{
natural_date(value)
}
\arguments{
\item{value}{}
}
\description{
Like naturalday, but will append a year for dates that are a year ago or
more.
}
