% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{natural_day}
\alias{natural_day}
\title{Natural Day}
\usage{
natural_day(value, format = "\%b \%d")
}
\arguments{
\item{format}{}
}
\description{
For date values that are tomorrow, today or yesterday compared to present day
returns representing string. Otherwise, returns a string formatted according
to \code{format}
}
