% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesize.R
\name{natural_size}
\alias{natural_size}
\title{Convert bytes to a more natural representation}
\usage{
natural_size(bytes, suffix_type = "decimal", fmt = "\%.1f")
}
\arguments{
\item{bytes}{Number of bytes}

\item{suffix_type}{One of 'decimal', 'binary', 'gnu'}

\item{fmt}{Extra number formatting}
}
\description{
Convert bytes to a more natural representation
}
\examples{
natural_size(3000)
}
