% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{natural_time}
\alias{natural_time}
\title{Given a datetime or a number of seconds, return a natraul representation of that resolution that makes sense}
\usage{
natural_time(value, future = FALSE, use_months = TRUE)
}
\description{
Given a datetime or a number of seconds, return a natraul representation of that resolution that makes sense
}
