% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{number_as_comma}
\alias{number_as_comma}
\title{Convert an number to a string with comma separation}
\usage{
number_as_comma(value)
}
\arguments{
\item{value}{A numeric}
}
\value{
A string with comma separation every three digits
}
\description{
Just a wrapper around \code{format} with defaults for full digits
}
\examples{
number_as_comma(1000)
number_as_comma(10000)
}
