% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assert_that}
\alias{reexports}
\alias{is.date}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{assertthat}{\code{\link[assertthat]{assert_that}}, \code{\link[assertthat]{is.date}}}
}}

