% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{natural_date}
\alias{natural_date}
\title{Natural Date}
\usage{
natural_date(value)
}
\arguments{
\item{value}{A Date value}
}
\description{
Like natural day, but will append a year for dates that are a year or
more in the past or future
}
\examples{
natural_date(Sys.Date())
natural_date(Sys.Date()-10)
}
\seealso{
\code{natural_day}
}
