% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/time.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assert_that}
\alias{reexports}
\alias{is.date}
\alias{reexports}
\alias{is.count}
\alias{reexports}
\alias{now}
\alias{reexports}
\alias{today}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{assertthat}{\code{\link[assertthat]{assert_that}}, \code{\link[assertthat]{is.date}}, \code{\link[assertthat]{is.count}}}

  \item{lubridate}{\code{\link[lubridate]{now}}, \code{\link[lubridate]{today}}}
}}

