% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{seconds_to_natural_delta}
\alias{seconds_to_natural_delta}
\title{Takes in a number of seconds and computes a "human" delta}
\usage{
seconds_to_natural_delta(seconds, use_months = TRUE)
}
\arguments{
\item{seconds}{A positive number of seconds}

\item{use_months}{Boolean whether we should (imprecisely) use months as a unit}
}
\description{
Takes in a number of seconds and computes a "human" delta
}
\seealso{
\code{natural_time}
}
